# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NotedThat HTTP API",
    "description": "The versioned machine API of a NotedThat server. Paths are relative to the server URL, `/api/v1`.\n\nThis document is generated from the server's code and is authoritative for paths, parameters, schemas and status codes. `docs/API.md` explains behaviour; where the two disagree, `docs/API.md` is wrong.\n\nEvery response carries `x-request-id`, echoing the request's when it sent one.",
    "contact": {
      "name": "NotedThat Contributors"
    },
    "license": {
      "name": "MPL-2.0",
      "identifier": "MPL-2.0"
    },
    "version": "0.12.3"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/knowledgebases/{kb_slug}/replace/{object_path}": {
      "post": {
        "tags": [
          "objects"
        ],
        "summary": "Replace text within an object.",
        "description": "Finds `old_string` in the object at `object_path` and replaces it with `new_string`, atomically against `If-Match`, which is required and must be a single entity tag. The object must be no larger than the server's patchable size limit.",
        "operationId": "replace_object",
        "parameters": [
          {
            "name": "kb_slug",
            "in": "path",
            "description": "The knowledge base.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KbSlug"
            }
          },
          {
            "name": "object_path",
            "in": "path",
            "description": "The object's path within the knowledge base, percent-encoded as one segment:\n`docs/rfc/7231.md` is sent as `docs%2Frfc%2F7231.md`. The server also accepts\nits `/` unencoded, which is how it writes `Location`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "The object's current entity tag. `*` and lists are refused `400`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-NotedThat-Source",
            "in": "header",
            "description": "`mcp` attributes the change events to the MCP server. Informational.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replaced.",
            "headers": {
              "Content-Location": {
                "schema": {
                  "type": "string"
                },
                "description": "The object's URL."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "The new entity tag."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableReplace"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/BackendUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AmbiguousMatchError": {
        "type": "object",
        "description": "The error envelope of a replace whose `old_string` occurs more than once.",
        "required": [
          "error",
          "message",
          "request_id",
          "match_count"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Always `ambiguous_match`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": "string",
            "description": "The request's id, the same value as the `x-request-id` response header."
          },
          "match_count": {
            "type": "integer",
            "format": "int64",
            "description": "How many times `old_string` occurs.",
            "minimum": 0
          }
        }
      },
...
      "Error": {
        "type": "object",
        "description": "The error envelope every JSON error response carries.",
        "required": [
          "error",
          "message",
          "request_id"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": "string",
            "description": "The request's id, the same value as the `x-request-id` response header."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable error code. Clients branch on this, never on `message`.",
        "enum": [
          "invalid_request",
          "malformed_range",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "gone",
          "payload_too_large",
          "precondition_failed",
          "no_match",
          "ambiguous_match",
          "precondition_required",
          "internal_error",
          "backend_unavailable",
          "request_timeout"
        ]
      },
...
      "KbSlug": {
        "type": "string",
        "description": "A knowledge base slug: 1 to 40 of `a-z`, `0-9` and `-`, not starting or ending with `-`.",
        "pattern": "^[a-z0-9]([a-z0-9-]{0,38}[a-z0-9])?$"
      },
...
      "Refusal": {
        "type": "object",
        "description": "The error envelope of a request the server refused before handling it: at its limit of requests in flight, or out of time.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "`backend_unavailable` or `request_timeout`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request's id, when one had been assigned."
          }
        }
      },
      "ReplaceBody": {
        "type": "object",
        "description": "A text replacement within one object.",
        "required": [
          "old_string",
          "new_string"
        ],
        "properties": {
          "old_string": {
            "type": "string",
            "description": "The text to find. Must not be empty."
          },
          "new_string": {
            "type": "string",
            "description": "The text to put in its place."
          },
          "replace_all": {
            "type": "boolean",
            "description": "Replace every occurrence. Otherwise exactly one must exist."
          }
        }
      },
      "ReplaceResponse": {
        "type": "object",
        "description": "The outcome of a replace.",
        "required": [
          "etag",
          "match_count",
          "total_bytes"
        ],
        "properties": {
          "etag": {
            "type": "string",
            "description": "The object's new entity tag."
          },
          "match_count": {
            "type": "integer",
            "format": "int64",
            "description": "How many occurrences were replaced.",
            "minimum": 0
          },
          "total_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "The object's new size in bytes.",
            "minimum": 0
          }
        }
      },
...
    },
    "responses": {
      "BackendUnavailable": {
        "description": "`backend_unavailable`: a backend or queue is unavailable or full, or the server is at its limit of requests in flight. Retry after `Retry-After`. A write answered this way may already be stored; repeat it.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            },
            "description": "Seconds to wait: `5`."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "$ref": "#/components/schemas/Refusal"
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "The request is malformed: `invalid_request` or `malformed_range`. A query string the server cannot parse at all is answered `400` with a plain-text body instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "`forbidden`: the credential is valid and the knowledge base's access policy does not grant this operation. An anonymous caller is answered `404` instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "`request_timeout`: the request did not complete within the server's timeout.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Refusal"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "`internal_error`: the server failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found`: no such knowledge base or object — or one an anonymous caller may not see, which reads identically.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "`payload_too_large`: the body exceeds the route's limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "`precondition_failed`: a conditional header's precondition did not hold.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "`precondition_required`: this write must be conditional; send `If-Match`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RequestTimeout": {
        "description": "`request_timeout`: the request body stopped arriving for longer than the server's idle limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Refusal"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "`unauthorized`: no credential where one is required, or one that did not verify.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "`Bearer resource_metadata=\"…\"`, when the deployment publishes OAuth protected resource metadata."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UnprocessableReplace": {
        "description": "`no_match`: `old_string` does not occur; or `ambiguous_match`: it occurs more than once and `replace_all` is false.",
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "$ref": "#/components/schemas/AmbiguousMatchError"
                }
              ]
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "service token or OIDC access token (JWT)",
        "description": "`Authorization: Bearer <token>`: the deployment's service token, or an access token from the configured OIDC issuer. A malformed or unverified credential is always `401`, never treated as anonymous."
      }
    }
  },
...
}